feat: paging, docker, auth, caching

+++ item/service.go
@@ -10,10 +10,10 @@ import (
 
 type ItemRepo interface {
 	Save(item *domain.ItemCreation) error
-	GetAll() ([]domain.Item, error)
-	GetByID(id uuid.UUID) (domain.Item, error)
-	Update(id uuid.UUID, item *domain.ItemUpdate) error
-	Delete(id uuid.UUID) error
+	GetAll(filter map[string]any, paging *clients.Paging) ([]domain.Item, error)
+	GetItem(filter map[string]any) (domain.Item, error)
+	Update(filter map[string]any, item *domain.ItemUpdate) error
+	Delete(filter map[string]any) error
 }
 
 type itemService struct {
@@ -39,8 +39,9 @@ func (s *itemService) CreateItem(item *domain.ItemCreation) error {
 	return nil
 }
 
-func (s *itemService) GetAllItem() ([]domain.Item, error) {
-	items, err := s.itemRepo.GetAll()
+func (s *itemService) GetAllItem(userID uuid.UUID, paging *clients.Paging) ([]domain.Item, error) {
+	filter := map[string]any{"user_id": userID}
+	items, err := s.itemRepo.GetAll(filter, paging)
 	if err != nil {
 		return nil, clients.ErrCannotListEntity(domain.Item{}.TableName(), err)
 	}
@@ -48,8 +49,8 @@ func (s *itemService) GetAllItem() ([]domain.Item, error) {
 	return items, nil
 }
 
-func (s *itemService) GetItemByID(id uuid.UUID) (domain.Item, error) {
-	item, err := s.itemRepo.GetByID(id)
+func (s *itemService) GetItemByID(id, userID uuid.UUID) (domain.Item, error) {
+	item, err := s.itemRepo.GetItem(map[string]any{"id": id, "user_id": userID})
 	if err != nil {
 		return domain.Item{}, clients.ErrCannotGetEntity(item.TableName(), err)
 	}
@@ -57,9 +58,9 @@ func (s *itemService) GetItemByID(id uuid.UUID) (domain.Item, error) {
 	return item, nil
 }
 
-func (s *itemService) UpdateItem(id uuid.UUID, item *domain.ItemUpdate) error {
+func (s *itemService) UpdateItem(id, userID uuid.UUID, item *domain.ItemUpdate) error {
 	item.UpdatedAt = time.Now()
-	err := s.itemRepo.Update(id, item)
+	err := s.itemRepo.Update(map[string]any{"id": id, "user_id": userID}, item)
 	if err != nil {
 		return clients.ErrCannotUpdateEntity(item.TableName(), err)
 	}
@@ -67,8 +68,8 @@ func (s *itemService) UpdateItem(id uuid.UUID, item *domain.ItemUpdate) error {
 	return nil
 }
 
-func (s *itemService) DeleteItem(id uuid.UUID) error {
-	err := s.itemRepo.Delete(id)
+func (s *itemService) DeleteItem(id, userID uuid.UUID) error {
+	err := s.itemRepo.Delete(map[string]any{"id": id, "user_id": userID})
 	if err != nil {
 		return clients.ErrCannotDeleteEntity(domain.Item{}.TableName(), err)
 	}
